+++ sccfm_core/__init__.py
@@ -2,6 +2,7 @@ from sccfm_core.errors import SccApiError
 from sccfm_core.factories.api_client_factory import ApiClientFactory
 from sccfm_core.services.health_service import HealthService, HealthStatus
 from sccfm_core.services.inventory import (
+    AsaBootImageService,
     AsaBootRegistryService,
     AsaCommandLineService,
     AsaDiskFileService,
@@ -14,6 +15,7 @@ from sccfm_core.services.inventory import (
 
 __all__ = [
     "ApiClientFactory",
+    "AsaBootImageService",
     "AsaBootRegistryService",
     "AsaCommandLineService",
     "AsaDiskFileService",